+++ Sources/SwiftOSCIONIO/TCP/Client/OSCTCPClientChannelHandler.swift
@@ -43,7 +43,7 @@ extension OSCTCPClientChannelHandler: ChannelInboundHandler {
 
         let remoteAddress = context.remoteAddress
         let remoteHost = remoteAddress?.ipAddress ?? ""
-        let remotePort = remoteAddress?.port ?? 0
+        let remotePort = UInt16(remoteAddress?.port ?? 0)
 
         oscServer._handle(receivedData: data, remoteHost: remoteHost, remotePort: remotePort)
     }
